Port and distance refactor part #1 BugzId: 2331

+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/com/mmxlabs/scheduler/optimiser/components/IXYPort.java
@@ -12,7 +12,7 @@ package com.mmxlabs.scheduler.optimiser.components;
  */
 public interface IXYPort extends IPort {
 
-	float getX();
+	double getX();
 
-	float getY();
+	double getY();
 }
